Updated Dietz model to use better syntax and extra functions
In addition some of the calculations is changed. Now there is explicitely an
f_inst and f_inst_lp version used, whereby the f_inst_lp is calculated after the
calc_f_inst function which alters the result a little bit.
In addition the signal coming out directly from the haircell processor is also
used to calculate the fine structure signals, which was not the case before!